+++ rice/cpp_api/Module_defn.hpp
@@ -1,7 +1,7 @@
 #ifndef Rice__Module_defn__hpp_
 #define Rice__Module_defn__hpp_
 
-#include "../detail/Exception_Handler_defn.hpp"
+#include "../detail/ExceptionHandler_defn.hpp"
 #include "../detail/ruby.hpp"
 #include "Object_defn.hpp"
 
@@ -44,41 +44,6 @@ namespace Rice
      */
     Class singleton_class() const;
 
-    //! Define an exception handler.
-    /*! Whenever an exception of type Exception_T is thrown from a
-     *  function defined on this class, functor will be called to
-     *  translate the exception into a ruby exception.
-     *  \param Exception_T a template parameter indicating the type of
-     *  exception to be translated.
-     *  \param functor a functor to be called to translate the exception
-     *  into a ruby exception.  This functor should re-throw the exception
-     *  as an Exception.
-     *  Example:
-     *  \code
-     *    class MyException : public std::exception { };
-     *    Data_Type<MyException> rb_cMyException;
-     *    Class rb_cFoo;
-     *
-     *    void translate_my_exception(MyException const & ex)
-     *    {
-     *      Data_Object<MyException> ex_(
-     *          new MyException(ex),
-     *          rb_cMyException);
-     *      throw Exception(ex_);
-     *    }
-     *
-     *    extern "C"
-     *    void Init_MyExtension()
-     *    {
-     *      rb_cMyException = define_class("MyException");
-     *      rb_cFoo = define_class("Foo")
-     *        .add_handler<MyException>(translate_my_exception);
-     *    }
-     *  \endcode
-     */
-    template<typename Exception_T, typename Functor_T>
-    Module& add_handler(Functor_T functor);
-
     //! Evaluate the given string in the context of the module.
     /*! This is equivalant to calling obj.module_eval(s) from inside the
      *  interpreter.
@@ -89,14 +54,8 @@ namespace Rice
     #include "shared_methods.hpp"
 
   protected:
-    std::shared_ptr<detail::Exception_Handler> handler() const;
-
     template<bool IsMethod, typename Function_T>
-    void wrap_native_call(VALUE klass, Identifier name, Function_T&& function,
-                          std::shared_ptr<detail::Exception_Handler> handler, MethodInfo* methodInfo);
-
-  private:
-    mutable std::shared_ptr<detail::Exception_Handler> handler_ = std::make_shared<Rice::detail::Default_Exception_Handler>();
+    void wrap_native_call(VALUE klass, Identifier name, Function_T&& function, MethodInfo* methodInfo);
   };
 
   //! Define a new module in the namespace given by module.