+++ rice/detail/Exception_Handler.hpp
@@ -1,8 +1,0 @@
-#ifndef Rice__detail__Exception_Handler__hpp_
-#define Rice__detail__Exception_Handler__hpp_
-
-#include "Exception_Handler_defn.hpp"
-#include "Exception_Handler.ipp"
-
-#endif // Rice__detail__Exception_Handler__hpp_
-