+++ test/test_Class.cpp
@@ -145,9 +145,10 @@ namespace
 
 TESTCASE(add_handler)
 {
+  register_handler<Silly_Exception>(handle_silly_exception);
+
   Class c(rb_cObject);
-  c.add_handler<Silly_Exception>(handle_silly_exception)
-    .define_function("foo", throw_silly_exception);
+  c.define_function("foo", throw_silly_exception);
 
   Object exc = detail::protect(rb_eval_string, "begin; foo; rescue Exception; $!; end");
   ASSERT_EQUAL(rb_eRuntimeError, CLASS_OF(exc));